category detaiils product fix
