+++ src/app/(commonUser)/categoryproduct/[slug]/page.jsx
@@ -1,8 +1,8 @@
-import ProductCategoryDetail from '@/components/pages/home/container/category/ProductCategoryDetail';
-import React from 'react';
+import ProductCategoryDetail from "@/components/pages/home/container/category/ProductCategoryDetail";
+import React from "react";
 
 const CategoryProductPage = async ({ params }) => {
-    const Params = await params;
+  const Params = await params;
   return (
     <div>
       <ProductCategoryDetail params={Params} />
@@ -11,5 +11,3 @@ const CategoryProductPage = async ({ params }) => {
 };
 
 export default CategoryProductPage;
-
-
